Refactor codes in the navbar
