+++ src/Components/NavBar/NavBar.js
@@ -1,8 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { GiHamburgerMenu } from 'react-icons/gi';
 import { AiOutlineClose } from 'react-icons/ai';
-import { motion, useAnimation } from 'framer-motion';
+import { motion } from 'framer-motion';
 import { NavLink } from 'react-router-dom';
+import UseNav from '../customHooks/UseNav';
 import logo from '../../assets/planet.png';
 import styles from './navBar.module.css';
 
@@ -12,53 +13,7 @@ const NavBar = () => {
   const handleScroll = () => {
     setOffsetY(window.scrollY);
   };
-  const animation = useAnimation();
-  const menuAnimation = useAnimation();
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    if (offsetY > 0) {
-      animation.start({
-        x: `${isOpen ? 0 : '100vw'}`,
-        transition: {
-          type: 'spring',
-          duration: 0.4,
-        },
-      });
-
-      menuAnimation.start({
-        x: 0,
-        position: 'fixed',
-        transition: {
-          type: 'spring',
-          stiffness: 100,
-          duration: 0.4,
-        },
-      });
-    }
-
-    if (offsetY === 0) {
-      animation.start({
-        x: 0,
-        transition: {
-          type: 'spring',
-          stiffness: 100,
-          duration: 0.4,
-        },
-      });
-
-      menuAnimation.start({
-        x: '30vw',
-        transition: {
-          type: 'spring',
-          duration: 0.4,
-        },
-      });
-    }
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [offsetY, isOpen]);
+  const { scrollAnimation, menuAnimation } = UseNav(offsetY, isOpen, handleScroll);
 
   const Links = [
     {
@@ -98,7 +53,7 @@ const NavBar = () => {
           </NavLink>
         </div>
       </div>
-      <motion.nav animate={animation} className={styles['nav-container']}>
+      <motion.nav animate={scrollAnimation} className={styles['nav-container']}>
         <ul
           className={
             isOpen
